The bug was that the flag “M_dataset.interpolated" that indicates if we need to redo the spatial interpolation was reset to “false" only when the mesh was adapted., whereas it should also be reset to “false" when we load new data.
For the Purely Lagrangian code, it is not really a problem, as we remesh very frequently, but for the pure Eulerian case, indeed it means that we always use the forcing of the first day (that explains the strange results I described this morning). We loaded the forcing correctly, but we simply didn’t use it as we didn’t do the interpolation.

That nasty bug was introduced by myself indeed, end of October, when I separated the loading of the data from the spatial interpolation.