Implementation of nextsim-ensemble
To perform ensemble simulations and data assimilation using nextsim,
a module called enkf is added. Scripts to run the model in ensemble mode
is in scripts/ensemble.

The enkf module includes a fortran library to perturb the atmospheric
forcing before interpolated to the neXtSIM mesh. The cpp classes and are
still under development but it works stand-alone in the meantime.

Ensemble scripts will submit an ensemble of neXtSIM simulations