Change name of thin ice class to young ice
We want to change the name of the "thin ice" class to "young ice". This
makes more sense, physically, and is more in line with the general
nomenclature (also WMO). Here we only change the name of variables (both
in the code and in output) - no physics or numerics changes.

Conflicts:
	model/finiteelement.cpp
	model/finiteelement.hpp
	model/gridoutput.hpp
	model/model_variable.hpp
	model/options.cpp