Update M_proc_mask every output time
We have to update M_proc_mask every output time step, Now that I'm
taking the grid displacement into account in the outputs.

It may be possible to do this more elegantly, by including the proc mask
in the list of variables M_elemental_variables. Problem is that the proc
mask must be interpolated using InterpFromMeshToMesh, but we cannot
guarantee that this is what will be done to M_element_variables - maybe
the user asked for ConservativeRemapping. If that's the case then the
code is getting too complicated. I can at least not see a simple
solution.