Remove M_proc_mask from GridOutput
Anton was right in that we don't need M_proc_mask in GridOutput. I've
been racking my brain as to why, because it was there for a reason.
Without it I saw a double-counting of velocities associated with
MPI-domain boundaries.

The only explanation I can find is that I switched from using a
Gmsh_mesh object to using a bamg object to describe the mesh. I guess
the bamg object doesn't include the ghost nodes, thus saving me the need
for M_proc_mask.