Remesh after reading restart if needed
This commit addresses issue #581, where OASIS would crash if the coupled
modelled is started from a restart where one (or more) of the triangle
angles is less than the critical angle (10 degrees, by default). In this
scenario the model asks OASIS for initial data, goes straight into the
remeshing and then asks for data again. OASIS doesn't like this second
request since it's being asked to give the same data twice and so brings
down the model.

Here we read in the restart file, check the angles and call the
``regrid`` function immediately if the angle is too small.