Add instantaneous divergence to mooring outputs
To accomplish this I added a diagnostic variable called D_divergence and
calculate the divergence in updateIceDiagnostics. This means that the
divergence is calculated once every large time step, using the
velocities from the end of the sub-cycling/short time stepping. In
theory we should calculate the divergence every sub-cycling/short time
step (for EVP and BBM) but the result is the same, since the element
shape is considered constant over a long time step (i.e. M_shape_coeff
stays constant during the sub-cycling).