Move getOptionFromMap and getAllowedOption to it's own class
I've moved the option handling functions getOptionFromMap and
getAllowedOption to it's own class called OptionHandler. They're
declared to be static, so they can be called also from the MeshHandler
class (and others). Works fine, except that I couldn't get the LOG()
mechanism to work. Seems like it relies on 'extern' variables, while my
OptionHandler relies on the functions being 'static' and that the two
don't mix. This needs to be fixed at a later date.