Change back to using pointers for gmshmesh, -geom, and -opt
I changed back to using pointers for the gmshmesh, -geom, and -opt
variables. This works fine - I just did a mistake with my previous
implementation. I also fixed some memory leaks related to the handling
of the bamg meshes. There's still a leak, but valgrind says it's because
of some code in libGmsh and the netCDF library.