More moving of mesh related things out of finiteelement.cpp
Most of the functions I move from FiniteElement into MeshHandler. The
exception is the ``measure`` function, which I moved into both GmshMesh
and GmshMeshSeq and ``jacobian``, which I moved into a MeshMath class.
This violates DRY - but GmshMesh and GmshMeshSeq already violate this
quite badly. Fixing that is an issue for a different commit/PR/issue.