Skip variables not in input files
In this commit I put a try-catch clause, so that the model doesn't crash
when reading an initialisation file that doesn't contain absolutely
everything we could want. It should for instance be possible to read in a
file with only ocean information or a file with both ice and ocean
information.