Address issue #615 using approach nr. 1 listed there
I went for approach 1 because the "prognostics" group is not properly
defined in the code. Turns out that the default output of
"Concentration", and "Thickness" is actually a diagnostic, because it's
the sum of M_conc and M_conc_young, etc. This mixes the prognostics
and diagnostics groups and in the end it was easier and quicker to
implement a per-variable output, like the moorings have it. But you can
still choose the bulk lists of "diagnostics" and "forcings". In fact,
you cannot choose from forcing variables, because they aren't
ModelVariable objects.

The default output is a list of variables, same as the default outputs
for the moorings. But we could want to either have the default output
"None" (which means only Time is output!) or the list Tim uses for
neXtSIM-F. This needs some discussion.