Change long and stdName of ocean_temp and ocean_salt
These outputs have ambiguous names and it's difficult to find a good one. In
the non-coupled case M_ocean_temp and M_ocean_salinity are the temperature and
salinity we're nudging the prognostic values (M_sst and M_sss) towards. In the
coupled case M_ocean_temp and M_ocean_salinity are the values received from the
coupler and are equal to M_sst and M_sss. I, therefore, propose to name them
"ocean temperature forcing" and "ocean salinity forcing". Not very good, but
hopefully comprehensible.