Add masking to the NEMO_icemod.nc input file reading
The NEMO_icemod.nc file is based on the CREG grid, which has a
coordinate singularity over land. We, therefore, need masking if we want
to do a coordinate based interpolation (as we do). Somehow earlier
versions of the code sometime managed to perform this interpolation -
even though that shouldn't work. But now it should always work.