Adds a check on MLD in checkFieldsFast
This adds a check on the MLD in checkFieldsFast, which throws an
std::runtime_error if MLD is outside [0, 10 km]. The lower bound is
quite strict (could also be a few mm) and the upper bound is just there
because the code requires an upper bound. In reality we just want to
check the lower bound.