Check MLD with assert
The check on MLD doesn't really belong in checkFieldsFast, because it's
not a prognostic field. We're just checking if we get crazy input, not
if neXtSIM is doing something wrong. This is better done with a simple
assert in the thermodynamics routine - as I've done in this commit.

This way is also better, because it's more like to be caught right away,
before NaNs spread all over the solution.