Fix bugs in ConservativeRemapping code
There was a strange bug where I forgot to wrap the counter (it should go
1 2 0, but went 1 2 3). The model should have crashed, but didn't.

Also, several (even stranger) bugs where Bamg returns doubles, but I
didn't convert them to integers. The conversion is done implicitly and
that should be ok, because all integers below 2^53 are exactly
representable as doubles. But apparently int(a - 1) does not give the
same results as int(a) - 1, even if a < 2^53.