Add an option to scale ocean drag with the Nansen number
This option (true by default) scales the ice-ocean drag so that when the
ice-atmosphere drag is modified by ridges the ice-ocean drag is scaled
so that the Nansen number stays the same. This does not take atmospheric
stability into account (and it should not). The handling of drag over
young ice is starting to be a bit inconsistent now, because the
ice-ocean drag under young ice is not taken into account in the
dynamics.